Flag handling

+++ src/messages/inter/op_codes.rs
@@ -7,7 +7,7 @@ pub enum OpCodes {
 
 impl OpCodes {
 
-    pub fn get_op_from_code(code: u16) -> Result<Self, String> {
+    pub fn get_op_from_code(code: u8) -> Result<Self, String> {
         for c in [Self::Query, Self::IQuery, Self::Status] {
             if c.get_code() == code {
                 return Ok(c);
@@ -17,7 +17,7 @@ impl OpCodes {
         Err(format!("Couldn't find for code: {}", code))
     }
 
-    pub fn get_code(&self) -> u16 {
+    pub fn get_code(&self) -> u8 {
         match self {
             Self::Query => 0,
             Self::IQuery => 1,
